feat: Introduce `redux`
Currently, we're only using it to store the current language code. This
is so we can avoid using `window.location` in the actual code, as this
hinders us from using server side rendering.

Technically, `redux` isn't needed for this, since this code won't ever
change wihin a SPA session (changing it requires reload, as it's a
different hostname). But we'll want `redux` for future things anyway.