fix(webpack): Make sure images are bundled well
This makes sure we bundle them in the same way on the client and on the
server.

I didn't want to mess around with a `.json` loader, as that needs to
work as it is for translations and settings, so I decided to remove the
manifest file for now. Not sure if it's really needed. We can add it
back later if so.